Move persistent_async_server fixture to conftest
